Update implemetation for collecting data. Add possibility to get Account WN8 increase data fo tanks

+++ src/main/java/com/vlasovartem/wotalyzer/controller/wn8/ExpectedDataController.java
@@ -30,7 +30,7 @@ public class ExpectedDataController {
     }
 
     @RequestMapping("/get/filtered")
-    public List<TankExpectedData> findByIds(@RequestParam List<Long> tankIds) {
+    public List<TankExpectedData> findByIds(@RequestParam List<Integer> tankIds) {
         return dataService.findByIds(tankIds);
     }
 }
